requirements.txt
streamlit
crewai
pandas
matplotlib
seaborn
langchain-google-genai
crewai-tools
litellm
google-generativeai